pyproject.toml
[project]
name = "bot-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.7.1",
    "dotenv>=0.9.9",
    "gradio>=6.9.0",
    "python-telegram-bot>=22.7",
    "requests>=2.32.5",
    "sentence-transformers>=5.3.0",
]
